Updates to makedoc; removed unused hooks directory; added documentation for boore2003 and beyer_bommer_2006.py

+++ makedocs.py
@@ -35,17 +35,20 @@ def main(args):
 
     # where should .rst files, Makefile, _build folder be written?
     SPHINX_DIR = os.path.join(os.path.expanduser('~'), '__api-doc')
+    print("Sphinx dir: %s" %SPHINX_DIR)
 
-    #where should the temporary clone of the shakemap gh-pages repo live?
+    # where should the temporary clone of the shakemap gh-pages repo live?
     CLONE_DIR=os.path.join(os.path.expanduser('~'),'__shake-doc','html')
-
+    print("Clone dir: %s" %CLONE_DIR)
     # where is the repository where this script is running?
     REPO_DIR = os.path.dirname(os.path.abspath(__file__))  # where is this script?
+    print("Repo dir: %s" %REPO_DIR)
 
     # get the human friendly version of the ShakeMap version
     res, verstr, stderr = getCommandOutput('git describe')
     if not len(verstr.strip()):
         verstr = DEFAULT_TAG
+    verstr = verstr.strip()
 
     # what is the package called and who are the authors
     PACKAGE = "shakemap"
@@ -59,21 +62,23 @@ def main(args):
     make_cmd = stdout.strip()
 
     try:
-        #clone the repository
-        sys.stderr.write('Cloning shakemap gh-pages branch...\n')
-        if os.path.isdir(CLONE_DIR):
-            shutil.rmtree(CLONE_DIR)
-        clonecmd = 'git clone -b gh-pages https://github.com/usgs/shakemap.git %s' % CLONE_DIR
-        res,stdout,stderr = getCommandOutput(clonecmd)
-        if not res:
-            raise Exception('Could not clone gh-pages branch.')
+        # clone the repository
+        if args.post:
+            sys.stderr.write('Cloning shakemap gh-pages branch...\n')
+            if os.path.isdir(CLONE_DIR):
+                shutil.rmtree(CLONE_DIR)
+            clonecmd = 'git clone -b gh-pages https://github.com/usgs/shakemap.git %s' % CLONE_DIR
+            res,stdout,stderr = getCommandOutput(clonecmd)
+            if not res:
+                raise Exception('Could not clone gh-pages branch.')
 
-        #remove the current apidocs folder from that repo
-        apidocfolder = os.path.join(CLONE_DIR,'apidoc') #this should probably not have 'html' in the path
+        # remove the current apidocs folder from that repo
+        apidocfolder = os.path.join(CLONE_DIR,'apidoc')
+        # this should probably not have 'html' in the path
         if os.path.isdir(apidocfolder):
             shutil.rmtree(apidocfolder)
 
-        #run the make command to build the shakemap manual
+        # run the make command to build the shakemap manual
         sys.stderr.write('Building shakemap manual (HTML)...\n')
         makefile = os.path.join(REPO_DIR,'doc','Makefile')
         os.chdir(os.path.join(REPO_DIR,'doc'))
@@ -82,23 +87,26 @@ def main(args):
         if not res:
             raise Exception('Could not build the ShakeMap manual - error "%s".' % stderr)
         
-        #run the sphinx api doc command
+        # run the sphinx api doc command
         sys.stderr.write('Building shakemap API documentation (REST)...\n')
         package_dir = os.path.join(REPO_DIR,'shakemap')
         sphinx_cmd = 'sphinx-apidoc -o %s -f -l -F -H %s -A "%s" -V %s %s' % (SPHINX_DIR,PACKAGE,AUTHORS,verstr,package_dir)
+        print("\n'%s'\n" %sphinx_cmd)
         res,stdout,stderr = getCommandOutput(sphinx_cmd)
         if not res:
             raise Exception('Could not build ShakeMap API documentation - error "%s".' % stderr)
 
-        #run the make command to build the shakemap manual (pdf version)
-        sys.stderr.write('Building shakemap manual (PDF)...\n')
-        os.chdir(os.path.join(REPO_DIR,'doc'))
-        manualcmd = '%s latexpdf' % make_cmd
-        res,stdout,stderr = getCommandOutput(manualcmd)
-        if not res:
-            raise Exception('Could not build the PDF version of the ShakeMap manual - error "%s".' % stderr)
+        # run the make command to build the shakemap manual (pdf version)
+        if not args.nopdf:
+            sys.stderr.write('Building shakemap manual (PDF)...\n')
+            os.chdir(os.path.join(REPO_DIR,'doc'))
+            manualcmd = '%s latexpdf' % make_cmd
+            res,stdout,stderr = getCommandOutput(manualcmd)
+            if not res:
+                raise Exception('Could not build the PDF version of the ShakeMap manual - error "%s".' % stderr)
 
-        #this has created a conf.py file and a Makefile.  We need to "edit" the conf.py file to include the ReadTheDocs theme.
+        # this has created a conf.py file and a Makefile.
+        # We need to "edit" the conf.py file to include the ReadTheDocs theme.
         fname = os.path.join(SPHINX_DIR,'conf.py')
         f = open(fname,'at')
         f.write("sys.path.insert(0, os.path.abspath('%s'))\n" % (REPO_DIR))
@@ -107,21 +115,21 @@ def main(args):
         f.write("html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]\n")
         f.close()
 
-        #Go to the sphinx directory and build the html
+        # Go to the sphinx directory and build the html
         sys.stderr.write('Building shakemap API documentation (HTML)...\n')
         os.chdir(SPHINX_DIR)
         res,stdout,stderr = getCommandOutput('%s html' % make_cmd)
         if not res:
             raise Exception('Could not build HTML for API documentation. - error "%s"' % stderr)
         
-        #copy the generated content to the gh-pages branch we created earlier
+        # copy the generated content to the gh-pages branch we created earlier
         htmldir = os.path.join(SPHINX_DIR,'_build','html')
         if os.path.isdir(apidocfolder):
             shutil.rmtree(apidocfolder)
         shutil.copytree(htmldir, apidocfolder)
 
         if args.post:
-            #cd to directory above where html content was pushed
+            # cd to directory above where html content was pushed
             htmldir = os.path.join(CLONE_DIR) #this should probably not have 'html' in it
             os.chdir(htmldir)
             res,stdout,stderr = getCommandOutput('touch .nojekyll')
@@ -155,6 +163,8 @@ if __name__ == '__main__':
                         help='Post content to the web')
     parser.add_argument('-c', '--clean', action='store_true', default=False,
                         help='Clean up local directories containing generated HTML content')
+    parser.add_argument('-n', '--nopdf', action='store_true', default=False,
+                        help='Suppress PDF building. ')
 
     pargs = parser.parse_args()
     main(pargs)
